+++ actions/requirements.txt
@@ -1,0 +1,7 @@
+python-dotenv
+google-api-python-client
+google-auth-httplib2
+google-auth-oauthlib
+python-dateutil
+dhooks
+pynacl
